✅ 챕터9. service discovery   - 인프라컨테이너 생성   - application property 분리   - 테스트코드 추가   - pinpoint 설정 테스트중

+++ backend/review-service/src/main/java/dev/greatseo/reviewservice/repository/ReviewEntity.java
@@ -13,7 +13,7 @@ public class ReviewEntity {
     private int id;
 
     @Version
-    private int version;
+    private Long version;
 
     private int productId;
     private int reviewId;
@@ -45,11 +45,11 @@ public class ReviewEntity {
         this.id = id;
     }
 
-    public int getVersion() {
+    public Long getVersion() {
         return version;
     }
 
-    public void setVersion(int version) {
+    public void setVersion(Long version) {
         this.version = version;
     }
